backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
python-multipart==0.0.26
anthropic==0.97.0
openai>=1.30
httpx>=0.28
google-auth>=2.30
requests>=2.32
aiosqlite==0.22.1
pillow==12.2.0
pdf2image==1.17.0
pdfplumber==0.11.9
pytesseract==0.3.13
